Move initial_db fixture to test_common.
This will be useful for other parts of the test suite.

Since initial_db uses specific extensions, we do the imports *in* the
function (after checking that they've already avoided). Otherwise,
they'd bleed into anything that imported test_common.
